requirements.txt
Brotli==1.0.9
click==8.1.3
colorama==0.4.5
dash==2.5.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-iconify==0.1.2
dash-mantine-components==0.10.2
dash-table==5.0.0
Flask==2.1.2
Flask-Compress==1.12
greenlet==1.1.2
gunicorn==20.1.0
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.23.0
pandas==1.4.3
plotly==5.9.0
psycopg2==2.9.3
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
six==1.16.0
SQLAlchemy==1.4.39
tenacity==8.0.1
Werkzeug==2.1.2
zipp==3.8.0
scipy~=1.8.1